\name{print.sienaBayesFit}
\alias{print.sienaBayesFit}
\alias{sienaBayes}
\title{Methods for processing sienaBayes objects}
\usage{
\method{print}{sienaBayesFit}(x, ...)
}
\description{
  \code{print} method for \code{sienaBayesFit} objects.
}
\arguments{
  \item{x}{An object of class \code{sienaBayesFit}}
  \item{\dots}{For extra arguments (none used at present)}
}
\value{
 The function \code{print.sienaBayesFit} prints results
 of a Bayesian analysis, as resulting from function \code{\link{bayes}}.
}
\seealso{\code{\link{bayes}}}
\references{
  See also \url{http://www.stats.ox.ac.uk/~snijders/siena/}
}
\author{Tom Snijders}
\keyword{ methods }
